Added the ability to actually play!
Even though right now the cpu isn't programmed to make intelligent
choices for which spot to pick, the user is now able to choose where to
mark their spot (the spots are displayed as numbers so the user knows
which number is associated with which position). Need to complete the
cpuCalc function which chooses where to mark the cpu's choice.