Woooohoooo (works!)
Okay so managed to get the logic down to get the cpu being a decent
player to play against. I've managed to win a couple of times and I've
lost lots of times. There's a bug where even though the user is the last
player to take a turn (for a tie or a win) the cpu still takes its turn.